requirements.txt
flask==3.0.0
groq>=0.13.0
aiohttp==3.9.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
numpy>=1.24.0

# 5-Person Team Dependencies
pydantic>=2.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
typer>=0.9.0
asyncpg>=0.29.0
alembic>=1.12.0
aiofiles>=23.0.0
httpx>=0.25.0
websockets>=11.0.0
mcp>=1.0.0
